Make it possible to customize image insertion code. (#20)
* Make it possible to customize image insertion code.

It can be useful to insert more that just a markdown image tag, for
example to provide additional control on the layout of images. Since
there's no convenient way to handle this on the JavaScript side, we
generate the code to insert in Python. Then it can be overridden with
the existing extensibility mechanism, that is, with a custom view.

* Restore backwards-compatibility.

The previous commit is backwards-incompatible for users who pointed
MARKDOWNX_UPLOAD_URLS_PATH to a custom view. This commit restores
backwards-compatibility with the previous API.