Defined a variable to determine the version.
Defined the variable `is_post_10` to establish the version, making the backward compatibility methods easier to read. The variable is a `bool`, and is `True` if Django is 1.11+. 
The variable facilitates the handling of the `renderer` method which was deprecated in Django 1.11, and further allows for the usage of alternative templates.